+++ src/event_connectors/actionnetwork.py
@@ -1,6 +1,6 @@
 import pyactionnetwork
 import requests
-from events import ActionNetworkEvent
+from event_models.events import ActionNetworkEvent
 
 CREATION_WINDOW_DAYS = 365
