+++ src/sync.py
@@ -8,9 +8,9 @@ from pprint import pprint
 
 import boto3
 
-from actionnetwork import ActionNetwork
-from airtable import Airtable
-from events import EventDiffer
+from event_connectors.actionnetwork import ActionNetwork
+from event_connectors.airtable import Airtable
+from event_models.events import EventDiffer
 
 SLACK_CHANNEL = os.environ['SLACK_CHANNEL']
 SLACK_FOOTER_URL = os.environ['SLACK_FOOTER_URL']
